Refactor coerce operators to inherit CScalarCoerceBase [#117186547]
