+++ libgpopt/include/gpopt/operators/CScalarArrayCoerceExpr.h
@@ -20,9 +20,7 @@
 #define GPOPT_CScalarArrayCoerceExpr_H
 
 #include "gpos/base.h"
-#include "gpopt/base/COptCtxt.h"
-#include "gpopt/operators/CScalar.h"
-#include "gpopt/base/CDrvdProp.h"
+#include "gpopt/operators/CScalarCoerceBase.h"
 
 namespace gpopt
 {
@@ -36,29 +34,16 @@ namespace gpopt
 	//		Scalar Array Coerce Expr operator
 	//
 	//---------------------------------------------------------------------------
-	class CScalarArrayCoerceExpr : public CScalar
+	class CScalarArrayCoerceExpr : public CScalarCoerceBase
 	{
 
 		private:
-
 			// catalog MDId of the element function
 			IMDId *m_pmdidElementFunc;
 		
-			// catalog MDId of the result type
-			IMDId *m_pmdidResultType;
-
-			// output type modifications
-			INT m_iMod;
-
 			// conversion semantics flag to pass to func
 			BOOL m_fIsExplicit;
 		
-			// coercion form
-			ECoercionForm m_ecf;
-
-			// location of token to be coerced
-			INT m_iLoc;
-
 			// private copy ctor
 			CScalarArrayCoerceExpr(const CScalarArrayCoerceExpr &);
 
@@ -81,7 +66,6 @@ namespace gpopt
 			~CScalarArrayCoerceExpr()
 			{
 				m_pmdidElementFunc->Release();
-				m_pmdidResultType->Release();
 			}
 		
 			// return metadata id of element coerce function
@@ -90,34 +74,9 @@ namespace gpopt
 				return m_pmdidElementFunc;
 			}
 
-			// the type of the scalar expression
-			virtual
-			IMDId *PmdidType() const
-			{
-				return m_pmdidResultType;
-			}
-
-			// return type modification
-			INT IMod() const
-			{
-				return m_iMod;
-			}
-		
 			BOOL FIsExplicit() const
 			{
 				return m_fIsExplicit;
-			}
-
-			// return coercion form
-			ECoercionForm Ecf() const
-			{
-				return m_ecf;
-			}
-
-			// return token location
-			INT ILoc() const
-			{
-				return m_iLoc;
 			}
 
 			virtual
@@ -136,25 +95,6 @@ namespace gpopt
 			// match function
 			virtual
 			BOOL FMatch(COperator *) const;
-
-			// sensitivity to order of inputs
-			virtual
-			BOOL FInputOrderSensitive() const
-			{
-				return false;
-			}
-
-			// return a copy of the operator with remapped columns
-			virtual
-			COperator *PopCopyWithRemappedColumns
-				(
-				IMemoryPool *, //pmp,
-				HMUlCr *, //phmulcr,
-				BOOL //fMustExist
-				)
-			{
-				return PopCopyDefault();
-			}
 
 			// conversion function
 			static
