+++ libgpopt/include/gpopt/operators/CScalarCoerceViaIO.h
@@ -21,9 +21,7 @@
 #define GPOPT_CScalarCoerceViaIO_H
 
 #include "gpos/base.h"
-#include "gpopt/base/COptCtxt.h"
-#include "gpopt/operators/CScalar.h"
-#include "gpopt/base/CDrvdProp.h"
+#include "gpopt/operators/CScalarCoerceBase.h"
 
 namespace gpopt
 {
@@ -37,23 +35,10 @@ namespace gpopt
 	//		Scalar CoerceViaIO operator
 	//
 	//---------------------------------------------------------------------------
-	class CScalarCoerceViaIO : public CScalar
+	class CScalarCoerceViaIO : public CScalarCoerceBase
 	{
 
 		private:
-
-			// catalog MDId of the result type
-			IMDId *m_pmdidResultType;
-
-			// output type modifications
-			INT m_iMod;
-
-			// coercion form
-			ECoercionForm m_ecf;
-
-			// location of token to be coerced
-			INT m_iLoc;
-
 			// private copy ctor
 			CScalarCoerceViaIO(const CScalarCoerceViaIO &);
 
@@ -73,32 +58,6 @@ namespace gpopt
 			virtual
 			~CScalarCoerceViaIO()
 			{
-				m_pmdidResultType->Release();
-			}
-
-			// the type of the scalar expression
-			virtual
-			IMDId *PmdidType() const
-			{
-				return m_pmdidResultType;
-			}
-
-			// return type modification
-			INT IMod() const
-			{
-				return m_iMod;
-			}
-
-			// return coercion form
-			ECoercionForm Ecf() const
-			{
-				return m_ecf;
-			}
-
-			// return token location
-			INT ILoc() const
-			{
-				return m_iLoc;
 			}
 
 			virtual
@@ -112,29 +71,6 @@ namespace gpopt
 			const CHAR *SzId() const
 			{
 				return "CScalarCoerceViaIO";
-			}
-
-			// match function
-			virtual
-			BOOL FMatch(COperator *) const;
-
-			// sensitivity to order of inputs
-			virtual
-			BOOL FInputOrderSensitive() const
-			{
-				return false;
-			}
-
-			// return a copy of the operator with remapped columns
-			virtual
-			COperator *PopCopyWithRemappedColumns
-				(
-				IMemoryPool *, //pmp,
-				HMUlCr *, //phmulcr,
-				BOOL //fMustExist
-				)
-			{
-				return PopCopyDefault();
 			}
 
 			// conversion function
