+++ libgpopt/src/operators/CScalarCoerceBase.cpp
@@ -1,12 +1,12 @@
 //---------------------------------------------------------------------------
 //	Greenplum Database
-//	Copyright (C) 2014 Pivotal Inc.
+//	Copyright (C) 2017 Pivotal Inc.
 //
 //	@filename:
-//		CScalarCoerceViaIO.cpp
+//		CScalarCoerceBase.cpp
 //
 //	@doc:
-//		Implementation of scalar CoerceViaIO operators
+//		Implementation of scalar coerce operator base class
 //
 //	@owner:
 //
@@ -16,7 +16,7 @@
 //---------------------------------------------------------------------------
 
 #include "gpos/base.h"
-#include "gpopt/operators/CScalarCoerceViaIO.h"
+#include "gpopt/operators/CScalarCoerceBase.h"
 
 using namespace gpopt;
 using namespace gpmd;
@@ -24,13 +24,13 @@ using namespace gpmd;
 
 //---------------------------------------------------------------------------
 //	@function:
-//		CScalarCoerceViaIO::CScalarCoerceViaIO
+//		CScalarCoerceBase::CScalarCoerceBase
 //
 //	@doc:
 //		Ctor
 //
 //---------------------------------------------------------------------------
-CScalarCoerceViaIO::CScalarCoerceViaIO
+CScalarCoerceBase::CScalarCoerceBase
 	(
 	IMemoryPool *pmp,
 	IMDId *pmdidType,
@@ -39,38 +39,45 @@ CScalarCoerceViaIO::CScalarCoerceViaIO
 	INT iLoc
 	)
 	:
-	CScalarCoerceBase(pmp, pmdidType, iMod, ecf, iLoc)
+	CScalar(pmp),
+	m_pmdidResultType(pmdidType),
+	m_iMod(iMod),
+	m_ecf(ecf),
+	m_iLoc(iLoc)
 {
+	GPOS_ASSERT(NULL != pmdidType);
+	GPOS_ASSERT(pmdidType->FValid());
+
 }
 
 
 //---------------------------------------------------------------------------
 //	@function:
-//		CScalarCoerceViaIO::FMatch
+//		CScalarCoerceBase::FMatch
 //
 //	@doc:
 //		Match function on operator level
 //
 //---------------------------------------------------------------------------
-//BOOL
-//CScalarCoerceViaIO::FMatch
-//	(
-//	COperator *pop
-//	)
-//	const
-//{
-//	if (pop->Eopid() == Eopid())
-//	{
-//		CScalarCoerceViaIO *popCoerce = CScalarCoerceViaIO::PopConvert(pop);
-//
-//		return popCoerce->PmdidType()->FEquals(m_pmdidResultType) &&
-//				popCoerce->IMod() == m_iMod &&
-//				popCoerce->Ecf() == m_ecf &&
-//				popCoerce->ILoc() == m_iLoc;
-//	}
-//
-//	return false;
-//}
+BOOL
+CScalarCoerceBase::FMatch
+	(
+	COperator *pop
+	)
+	const
+{
+	if (pop->Eopid() == Eopid())
+	{
+		CScalarCoerceBase *popCoerce = dynamic_cast<CScalarCoerceBase*>(pop);
+
+		return popCoerce->PmdidType()->FEquals(m_pmdidResultType) &&
+				popCoerce->IMod() == m_iMod &&
+				popCoerce->Ecf() == m_ecf &&
+				popCoerce->ILoc() == m_iLoc;
+	}
+
+	return false;
+}
 
 
 // EOF
